Put ConnectivityAnalysisResult back in DcSensitivityAnalysis
To avoid conflicts between PRs
